Support for audio labels in tracks
We can now associate a label to each sound which gives better control to manage playback. We can now set a sound to loop and give it a label which can then used to stop it.

To demostrate this patch adds a sound effect when characters meditate: during meditation a small effect could be heard and this is using either channel left/right depending on the character position.