Fixed OS language detection
Fixed OS language detection

The language is chosen based on the value of the field Localization in Recursos\OUTPUTConfiguracion.ini.

If the value is absent the game will use the same language as the OS and save the setting to Configuracion.ini.

Refactor and clean up of the code to detect OS language

Deleted unused function GetLocaleLanguage()